requirements.txt
contourpy==1.2.0
cycler==0.12.1
flake8==7.0.0
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
mccabe==0.7.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
